rename to grid area ids
