+++ source/geh_common/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geh_common"
-version = "7.5.1"
+version = "7.5.2"
 description = "Functionality common to DataHub3 subsystems"
 readme = "README.md"
 requires-python = ">=3.11"
