+++ source/geh_common/src/geh_common/application/__init__.py
@@ -1,8 +1,8 @@
 from geh_common.application import settings
 from geh_common.application.retry_policy import retry_policy
 from geh_common.application.types.energy_supplier_ids import EnergySupplierIds
-from geh_common.application.types.grid_area_codes import GridAreaCodes
+from geh_common.application.types.grid_area_ids import GridAreaIds
 from geh_common.application.types.metering_point_ids import MeteringPointIds
 from geh_common.application.types.metering_point_types import MeteringPointTypes
 
-__all__ = ["settings", "GridAreaCodes", "EnergySupplierIds", "MeteringPointTypes", "MeteringPointIds", "retry_policy"]
+__all__ = ["settings", "GridAreaIds", "EnergySupplierIds", "MeteringPointTypes", "MeteringPointIds", "retry_policy"]
